requirements.txt
torch == 1.7.0
pytorch-ignite == 0.4.2
transformers ~= 3.0
datasets == 1.1.3
tensorboardX == 1.8
spacy

# dgl == 0.5.3